requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
python-decouple
drf-yasg
celery
redis
requests
django-cors-headers